fix: split file for better readability, adjust test and functions
