+++ package.json
@@ -126,7 +126,7 @@
         "@mui/icons-material": "^6.4.3",
         "@mui/material": "^6.4.3",
         "@mui/x-date-pickers": "^7.26.0",
-        "@react-pdf/renderer": "^3.1.11 || ^4.0.0",
+        "@react-pdf/renderer": "^4.4.1",
         "@sentry/react": "^8.54.0",
         "@stripe/react-stripe-js": "^5.4.1",
         "@stripe/stripe-js": "^8.5.3",
